+++ Content/NPCs/Enemy/ThroughChapter4/OriginiumSlug.png
