+++ Content/NPCs/Enemy/ThroughChapter4/OriginiumSlugAlpha.png
