+++ Content/NPCs/Enemy/ThroughChapter4/Soldier.png
